+++ src/assets/icons/exode-icon.tsx
