+++ src/components/country-card/country-card-languages-list.tsx
@@ -1,4 +1,4 @@
-import { LanguageType } from '../models/countries-types';
+import { LanguageType } from '../../models/countries-types';
 
 interface CountryCardLanguagesListPropsI {
   data: LanguageType[];
