+++ src/components/header/css/header.css
@@ -3,7 +3,3 @@
   justify-content: center;
   height: 75px; 
 }
-
-.header__toolbar {
- padding: 0 !important;
-}
