+++ src/constants/constants.ts
@@ -1,3 +1,3 @@
 export const labels = {
-  searchInput: 'Enter the country code to find it'
+  searchInput: 'Country code...'
 };
